requirements.txt
pillow==10.0.0
streamlit==1.27.2
pdf2image==1.16.3
PyMuPDF==1.22.3
requests==2.31.0
PyPDF2==3.0.1 
python-docx==0.8.11
opencv-python==4.10.0.84
langchain
langchain_community
langchain_google_genai
python-dotenv
pdf2image==1.16.3
pdfplumber